Big bad refatco
- remove picking recording mode
- split voice and playback
- don't even load the AI files when AI is not needed, for faster loading
