+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "raspilunii"
+name = "raspitalk"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
