+++ src/states.py
@@ -21,8 +21,8 @@ class WORKING_MODE(Enum):
     CONVERSATION_MODE = 0
     STORY_MODE = 1
     RANDOM_RECORDING_MODE = 2
-    PICK_RECORDING_MODE = 3
-    LAST = 4
+    # PICK_RECORDING_MODE = 3
+    LAST = 3
 
 
 class WORKING_LANGUAGE(Enum):
@@ -55,7 +55,9 @@ class InputControllerStateMachine:
         self.is_ai_available = is_ai_available
         if not self.is_ai_available:
             self.working_mode = WORKING_MODE.RANDOM_RECORDING_MODE
-        logging.info("InputControllerStateMachine initialized.")
+        logging.info(
+            f"InputControllerStateMachine initialized in mode {self.working_mode}."
+        )
 
     def next_state(self, input_event: INPUT_CONTROLLER_ACTION):
         """
@@ -116,13 +118,17 @@ class InputControllerStateMachine:
 
         elif input_event == INPUT_CONTROLLER_ACTION.MIDDLE_BUTTON_TOGGLE:
             # this will return the DISPLAY MODE
-            logging.info("Transitioning state on MIDDLE_BUTTON_TOGGLE")
+            logging.info(
+                f"Transitioning state from {self.menu_state} on MIDDLE_BUTTON_TOGGLE"
+            )
             # swich visual mode
             self.display_mode = DISPLAY_MODE(not bool(int(self.display_mode.value)))
             return self.display_mode
 
         elif input_event == INPUT_CONTROLLER_ACTION.RIGHT_BUTTON_TOGGLE:
-            logging.info("Transitioning state on RIGHT_BUTTON_TOGGLE")
+            logging.info(
+                f"Transitioning state from {self.menu_state} and {self.working_mode} on RIGHT_BUTTON_TOGGLE"
+            )
             # this will return the MENU_STATE
             if self.menu_state == MENU_STATE.MODE_CHOICE:
                 if self.working_mode == WORKING_MODE.STORY_MODE:
@@ -159,7 +165,9 @@ class InputControllerStateMachine:
             return self.menu_state
 
         elif input_event == INPUT_CONTROLLER_ACTION.LEFT_BUTTON_HELD:
-            logging.info("Transitioning state on LEFT_BUTTON_HELD")
+            logging.info(
+                f"Transitioning state from {self.menu_state} on LEFT_BUTTON_HELD"
+            )
             self.menu_state = MENU_STATE.MODE_CHOICE
             return self.menu_state
